move Compat methods back into LocaleUtils
