+++ gto-support-util/src/main/kotlin/org/ccci/gto/android/common/util/LocaleUtils.kt
@@ -1,10 +1,17 @@
 package org.ccci.gto.android.common.util
 
+import android.icu.util.ULocale
+import android.os.Build
+import androidx.annotation.RequiresApi
 import java.util.Locale
 
 object LocaleUtils {
-    // define a few fixed fallbacks
-    internal val FALLBACKS = mutableMapOf(
+    private val COMPAT = when {
+        Build.VERSION.SDK_INT < Build.VERSION_CODES.N -> Compat.Base()
+        else -> Compat.Nougat()
+    }
+
+    private val FIXED_FALLBACKS = mutableMapOf(
         // macro language fallbacks https://www.iana.org/assignments/language-subtag-registry/language-subtag-registry
         // Malagasy macro language
         "bhr" to "mg",
@@ -29,21 +36,68 @@ object LocaleUtils {
     // region Language fallback methods
     @JvmStatic
     fun addFallback(locale: String, fallback: String) {
-        require(FALLBACKS[locale] == null) { "$locale already has a fallback language defined" }
-        FALLBACKS[locale] = fallback
+        require(FIXED_FALLBACKS[locale] == null) { "$locale already has a fallback language defined" }
+        FIXED_FALLBACKS[locale] = fallback
     }
 
+    internal fun generateFallbacksSequence(locale: Locale): Sequence<Locale> = COMPAT.generateFallbacksSequence(locale)
+
     @JvmStatic
-    fun getFallback(locale: Locale) = LOCALE_COMPAT.generateFallbacksSequence(locale).firstOrNull()
+    fun getFallback(locale: Locale) = COMPAT.generateFallbacksSequence(locale).firstOrNull()
 
     @JvmStatic
     fun getFallbacks(locale: Locale) =
-        (sequenceOf(locale) + LOCALE_COMPAT.generateFallbacksSequence(locale)).distinct().toList().toTypedArray()
+        (sequenceOf(locale) + COMPAT.generateFallbacksSequence(locale)).distinct().toList().toTypedArray()
 
     @JvmStatic
     fun getFallbacks(vararg locales: Locale) = locales.asSequence()
-        .flatMap { sequenceOf(it) + LOCALE_COMPAT.generateFallbacksSequence(it) }
+        .flatMap { sequenceOf(it) + COMPAT.generateFallbacksSequence(it) }
         .distinct()
         .toList().toTypedArray()
     // endregion Language fallback methods
+
+    internal sealed interface Compat {
+        fun generateFallbacksSequence(locale: Locale): Sequence<Locale>
+
+        open class Base : Compat {
+            override fun generateFallbacksSequence(locale: Locale): Sequence<Locale> {
+                val builder = Locale.Builder().setLocaleSafe(locale)
+                return generateSequence {
+                    val currentLocale = builder.buildOrNull() ?: return@generateSequence null
+                    val directFallback = FIXED_FALLBACKS[currentLocale.toLanguageTag()]
+                    when {
+                        directFallback != null -> {
+                            val fallbackLocale = Locale.forLanguageTag(directFallback)
+                            builder.setLocaleSafe(fallbackLocale)
+                            fallbackLocale
+                        }
+                        currentLocale.extensionKeys.isNotEmpty() -> builder.clearExtensions().build()
+                        currentLocale.variant.isNotEmpty() -> builder.setVariant(null).build()
+                        currentLocale.country.isNotEmpty() -> builder.setRegion(null).build()
+                        currentLocale.script.isNotEmpty() -> builder.setScript(null).build()
+                        else -> {
+                            builder.clear()
+                            null
+                        }
+                    }
+                }.distinct()
+            }
+        }
+
+        @RequiresApi(Build.VERSION_CODES.N)
+        class Nougat : Base() {
+            override fun generateFallbacksSequence(locale: Locale) =
+                generateSequence(ULocale.forLocale(locale)) {
+                    val fixed = FIXED_FALLBACKS[it.toLanguageTag()]
+                    when {
+                        // fixed fallback
+                        fixed != null -> ULocale.forLanguageTag(fixed)
+                        // remove extensions as the fallback if any are defined
+                        locale.extensionKeys.isNotEmpty() -> ULocale.Builder().setLocale(it).clearExtensions().build()
+                        // use normal fallback behavior
+                        else -> it.fallback.takeUnless { it == ULocale.ROOT }
+                    }
+                }.drop(1).map { it.toLocale() }.distinct()
+        }
+    }
 }
